cambios

+++ src/repositories/cotizacion.repository.js
@@ -1,4 +1,4 @@
-const { Cotizacion } = require("../models");
+const { Cotizacion, Estado } = require("../models");
 
 exports.getAllCotizaciones = async () => {
     return await Cotizacion.findAll();
@@ -18,12 +18,26 @@ exports.createCotizacion = async (cotizacion) => {
 
 exports.updateCotizacion = async (id, cotizacion) => {
     return await Cotizacion.update(cotizacion, { where: { id } });
-}
+};
 
 exports.statusCotizacion = async (id) => {
     return await Cotizacion.update({ estado: false }, { where: { id } });
 }
 
 exports.deleteCotizacion = async (id) => {
+    const cotizacion = await Cotizacion.findOne({
+        where: { id }
+    });
+    
+    if (!cotizacion) {
+        throw new Error("Cotización no encontrada");
+    }
+
+    // const existePermiso = cotizacion.pedidoId && cotizacion.pedidoId.length > 0;
+    // const existeEstado = await Estado.findOne({ where: { id: cotizacion.estadoId } });
+    
+    // if (existePermiso || existeEstado) {
+    //     throw new Error("No se puede eliminar la cotización porque está asociada a registros en otras tablas");
+    // }
     return await Cotizacion.destroy({ where: { id } });
 }
